LayoutTests/http/tests/storage/setItem-and-reload-expected.txt

CONSOLE MESSAGE: line 25: Reloading...
Test that setting local storage, reloading and getting local storage works.

Reloaded
